New sketches - CPU with jumps, counter, register
I made a better counter, advcounter, that supports Start/Stop and Load.
Using that, I build a little CPU that supports jump, branch if equal and branch if not equal.
It's programmed to decrement a register by two until it's zero.
The new register, 8BitProReg, is an 8-Bit register that supports shifting. It's controlled with three
control inputs. The commands are: 0 - nothing, 1 - load, 2 - all to zero, 3 - shift left, 4 - shift right, 5 - ring shift left, 6 - ring shift right, 7 - all to one.