Preview works again (mostly), structural changes/additions
Outsourced preview to its own class. Now all components have added a method to show a preview, previewSymbol is now the passed component for the preview, so now previewSymbol just delegates the showPreview back to the
corresponding object. Still clicking twice on same basic component button or on other buttons needs to be fixed.